+++ src/core/common/rotr.cpp
@@ -5,7 +5,7 @@ uint32_t rotl32(uint32_t n, unsigned int c)
   const unsigned int mask = 0x1F;
 
   c &= mask;
-  return (n<<c) | (n>>( (-c)&mask ));
+  return (n<<c) | (n>>( (-(int)c)&mask ));
 }
 
 uint32_t rotr32(uint32_t n, unsigned int c)
@@ -13,7 +13,7 @@ uint32_t rotr32(uint32_t n, unsigned int c)
     const unsigned int mask = 0x1F;
     c &= mask;
 
-    uint32_t result = (n>>c) | (n<<( (-c)&mask ));
+    uint32_t result = (n>>c) | (n<<( (-(int)c)&mask ));
 
     return result;
 }
